+++ plugins/modules/key_provider.py
@@ -55,8 +55,10 @@ notes:
 seealso:
 - module: virtru.dsp_tructl.kas_registry
 - module: virtru.dsp_tructl.kas_key
+extends_documentation_fragment:
+- virtru.dsp_tructl.tructl
 author:
-- Virtru DSP Team
+- Virtru (@virtru)
 """
 
 EXAMPLES = r"""