+++ plugins/modules/namespace_key.py
@@ -45,8 +45,10 @@ notes:
 seealso:
 - module: virtru.dsp_tructl.namespace
 - module: virtru.dsp_tructl.kas_key
+extends_documentation_fragment:
+- virtru.dsp_tructl.tructl
 author:
-- Virtru DSP Team
+- Virtru (@virtru)
 """
 
 EXAMPLES = r"""