+++ plugins/modules/policy_export.py
@@ -65,7 +65,7 @@ seealso:
 - module: virtru.dsp_tructl.provision
 - module: virtru.dsp_tructl.auth
 author:
-- Virtru DSP Team
+- Virtru (@virtru)
 """
 
 EXAMPLES = r"""