+++ plugins/modules/resource_mapping.py
@@ -53,8 +53,10 @@ notes:
 seealso:
 - module: virtru.dsp_tructl.attribute
 - module: virtru.dsp_tructl.subject_mapping
+extends_documentation_fragment:
+- virtru.dsp_tructl.tructl
 author:
-- Virtru DSP Team
+- Virtru (@virtru)
 """
 
 EXAMPLES = r"""