+++ plugins/modules/resource_mapping_group.py
@@ -51,8 +51,10 @@ notes:
 seealso:
 - module: virtru.dsp_tructl.resource_mapping
 - module: virtru.dsp_tructl.namespace
+extends_documentation_fragment:
+- virtru.dsp_tructl.tructl
 author:
-- Virtru DSP Team
+- Virtru (@virtru)
 """
 
 EXAMPLES = r"""