+++ plugins/modules/subject_condition_set.py
@@ -50,8 +50,10 @@ notes:
 - Check mode is supported for all states.
 seealso:
 - module: virtru.dsp_tructl.subject_mapping
+extends_documentation_fragment:
+- virtru.dsp_tructl.tructl
 author:
-- Virtru DSP Team
+- Virtru (@virtru)
 """
 
 EXAMPLES = r"""